add synchronized to print methods
when running the program over a lot of files, not on the same file, but
it happened reliably, a few things wouldnt be printed and messed up the
formatting. synchronized seems to fix this. i think the issue is
`append` would be invoked in between the lines in `handle`. so the
buffer would be printed, then added to, but then wiped.